+++ R/suggest_orch_frequency.R
@@ -9,6 +9,8 @@
 #' Note this function is intended to be used interactively when deciding how often to
 #' schedule the orchestrator. Programmatic use is not recommended.
 #'
+#' **Deprecated.** This function is no longer actively maintained and may be removed
+#' in a future release.
 #' @param schedule MaestroSchedule object created by `build_schedule()`
 #' @inheritParams get_pipeline_run_sequence
 #'
@@ -16,8 +18,6 @@
 #' @export
 #'
 #' @keywords internal
-#'
-#' @deprecated
 #'
 #' @examples
 #'